[assembly] Remove image assembly comparison safety net
The compare_image_assembly_configs() safety net has been growing less
useful, as more exceptions are needing to be added to it.  Moving
forward, as packages are moved from the legacy bundle to the common
platform bundles, and as more structured configuration is added to the
system, this is a less and less useful safety net, and a before vs.
after comparison is needed instead (but cannot be expressed in a
single build run, and so must be performed outside of the build
system).

Change-Id: I52445cc8196326515683aeb22f4cf641ac9baa4c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705582
Commit-Queue: Aidan Wolter <awolter@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Shai Barack <shayba@google.com>