[scrutiny][verifiers] Remove /core/universe-resolver allowlist
Now that full-resolver-env uses /core/universe-resolver to resolve
fuchsia-pkg URLs (because /core/full-resolver was deleted), this
instance of the allowlist has largely become a change detector for
children of /core. The other instances of this allowlist are not
changed and still restrict the use of /core/universe-resolver.

Bug: 105528
Change-Id: I0b085a158e5fd5504979489357f34903dbf92562
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707304
Reviewed-by: Allison Pearce <ampearce@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>