[pci] Change policy on when we disable devices
It's possible that a device is enabled by the bootloader, or in use by
the kernel. If we probe the device and are unable to allocate an
existing window for it then we will now just leave it alone and no
longer expose any bind points for drivers. To enable this some changes
were made to cut down on the number of command register checks that were
needed, and an RAII Io disabler has been introduced to ensure the state
before and after our probing is coherent. Further improvements have been
made to cut down on command register modifications at the same time.

This should resolve issues on systems where the kernel carves out space
for something like a UART and the PCI driver is not able to maintain
that allocation with its own bookkeeping.

Fixed: 96645
Change-Id: I56c2c6bae1e90a7e5aa62773112c2cba13ca0fc6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679687
Reviewed-by: Simon Shields <simonshields@google.com>
Commit-Queue: Christopher Anderson <cja@google.com>