[build] Allow specifying image name
The image name is use as a selector in a metadata file. Before,
the name was presumed to be "kernel" which is correct almost
always.

In some cases, we use a modified zbi image, which has a custom image
name.

If the image name is not set, will default to he label name. Which
matches zbi_tests at this moment. The default for kernel image,
is "kernel" hardcoded.

Test: fx build
Change-Id: Ic78a32b849eff48e82ad19297b2a21683631cbf3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705262
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>