[virtualization] Allow some gpu/input queues to be optional.
The zircon drivers don't configure all the queues. With the virtio-gpu
we don't configure the cursorq and with the input device we don't
configure the statusq. The `take_stream` method will return `None` for
these queue when running against a zircon guest so we need to teach the
rest of the device to handle this situation.

Bug: 104229, 102870
Test: fx test device_tests
Change-Id: Ie16ace49ebcdf2840e225fe28412eb689dc33ea8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705532
Reviewed-by: David Hastings <dahastin@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>