[blobfs][host] Parallelize chunked compression
Add a new chunked compressor that uses a thread pool to compress the
chunks of the chunked compression format in parallel.

The blobfs host tool already compresses blobs in parallel but each blob
was being processed in a single thread. For builds with a few large
blobs, the entire build would be waiting for a couple threads to
compress these large blobs. The MultithreadedChunkedCompressor allows
the blobfs host tool to use all of threads on the system to compress
these large blobs which significantly reduces the build time of blobfs
images.

Bug: 105094
Change-Id: I22ad982ca72e7002bd773ce34570724d24fe809e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705682
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>