[fdio] Clean up svc fdio_ns_open calls.
There are several invocations of fdio apis that are used to open a
namespace directory.

These invocations extend the path provided with a `.`, and some have
comments mentioning that without extending the path, the open will fail.

These likely outdated changes to fdio_namespace::Connect in which a path
is extended with "." if the path was otherwise entirely within the
namespace.

These extensions, while not incorrect, should no longer be needed.

Bug: 105746
Change-Id: I524009e219b1e6861e164702f208ceb45e31c0d3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707549
Reviewed-by: Tamir Duberstein <tamird@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Luke Nicholson <lukenicholson@google.com>