[component][fuzzing] Add fuzz-test-runner
This CL adds a fuzz-test-runner as part of the component testing
framework. This runner acts almost identically to the elf-test-runner,
except that it additionally creates a channel, installs one end in the
started fuzz test, and passes the other to the fuzz-manager for use with
the fuchsia.fuzzer.Controller protocol.

Incidentally, it silences a warning from the transceiver if the
Controller client decides it doesn't care about receiving data it
requested and drops the connection (as happens in the fuzz-test-runner
tests).

Bug: 24740
Change-Id: Ib16da9d0d75f5a2cadb0ffdcea69f4221a3b76d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/558392
Reviewed-by: Cameron Finucane <eep@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>