[ffx][isolate] Allow overriding ascendd socket
The isolate library configuring the socket through the environment
variable prevents users of it from overriding the socket path, as is
done in the daemon self-test, causing it to not work correctly (though
apparently only some of the time).

Setting it both ways is no longer necessary, ASCENDD and overnet.socket
have a priority ordering and it is sufficient to set it as a default
configuration that can be overridden with command line arguments now.

Bug: 104562

Change-Id: Iec1d50c8eb1cea3d8d2dd8b0a6a1f8c95bcfae3b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708102
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>