[a11y][gesture-recognizers] Rename timing constants
Rename gesture recognizer timing constants in preparation for a
refactor. As of this CL, any constant that is defined in multiple
places will have the same name everywhere it is defined.

The next CL will put these constants into a separate header file, so
that we're not re-defining these values in multiple places.

Bug: 70569
Test: fx test //src/ui/a11y/lib:a11y-lib-tests
Change-Id: Ib4dd45d52fcf36c3672291f7945b19d67dc1db01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707008
Reviewed-by: Alexander Brusher <abrusher@google.com>
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Commit-Queue: Kevan Hollbach <khollbach@google.com>