[ffx] Route scheduling config from ffx test to test_manager
This changes ffx test and the run_test_suite library, which is used by ffx test to
run tests via test_manager. This library now takes in scheduling options
passed in through ffx test flags, and proceeds to invoke the
WithSchedulingOptions method in the RunBuilder FIDL protocol if the
client specified scheduling options. Since the run_test_suite library
now calls the WithSchedulingOptions method, test_manager needed to be
updated to serve the method call. Without a server implemention of the
WithSchedulingOptions in test_manager, calling this method from the
run_test_suite library resulted in an error. This change to test_manager
only sets up the use of scheduling options from ffx test, but it does
nothing with them at the moment. Note that there was also a change in
RunParams required by an API change in CL: 707022. This change allows
users to set the maximum number of parallel suites to be 0, indicating
that the server will choose a default value.

Test: ffx test run fuchsia-pkg://fuchsia.com/test_manager_test#meta/test_manager_test.cm --experimental-parallel-execution 10
Added request_scheduling_options_test to fx test run-test-suite-unit-tests
Bug: 105640

Change-Id: I1cec86fb3ef6beb11dad90f4367fd961a25515c8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707344
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Commit-Queue: Ismael Castro <iicastro@google.com>