[build] In-tree C++ builds at "HEAD" API level.
In order for C++ libraries in the SDK to make use of pending FIDL APIs
when built in-tree, it must be possible for the C++ API to mark its
availability to match.

clang does not have a concept of a HEAD API level, so here we emulate
it using a very large integer.

Bug: 104513
Change-Id: I055a4956029e9a045e5233412b026a39125fc8bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708010
Reviewed-by: Adam Barth <abarth@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>
API-Review: Adam Barth <abarth@google.com>
Commit-Queue: Tamir Duberstein <tamird@google.com>