[ffx] Add some advanced tuntap checks
Previously, the tun/tap networking mode validation was limited to "is an
interface listed on the system". No verification that it was the correct
interface, or that it was in a usable state was performed.

This CL reworks the "tap_available" utility method to glean more
information from the call to "ip". It will now look for the interface of
the appropriate name, verify that it is in fact a Tun/Tap interface,
that it is enabled, and that isn't being used by another process. The
error messages from the method provide context and remedial actions the
user can take, where available.

There are now two distinct front-end functions to call for this feature:
tap_available() checks that the interface is present and correctly
configured and unused, while tap_ready() also checks whether it's in the
"UP" state. tap_available() is meant for resolving the --net auto flag,
while tap_ready() is meant for network validation.

Bug: 103180
Bug: 103725

Change-Id: I54c86494eadfcf4f5ef25aeb9ba6eb30261132ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705304
Commit-Queue: Daniel Dougherty <doughertyda@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>