[build][testing] Extend `golden_file()` to `golden_files()`
This change introduces a version of `golden_file()` for multiple golden
file comparisons: `golden_files()`. While it is true that
`golden_file()` could just be invoked multiple times, it seems nicer to
save on executing more actions in the build (especially for minimal
added complexity). `golden_file()` is preserved as a more convenient,
singleton version but is made to use `golden_files()` under the hood.

This change is being introduced in the context of a new FIDL backend
that will produce multiple files to be checked against goldens.

Bug: 51002
Change-Id: Id0a02de8fac7432d0bb707998698ac729d0c5702
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707991
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Adam Perry <adamperry@google.com>