[bt][profiles] Migrate bt-a2dp-looback to CFv2
Define v2 manifests for the loopback test and A2DP sink/source
components.

Migrate the loopback shell package to a Fuchsia test. This test can then
be invoked by SL4F to start streaming.

Remove the optional argument to specify a streaming duration. This is a
fixed value specified in the E2E test code and is never used otherwise.

Update test_manager to provide the SessionAudioConsumerFactory
capability to system tests. This allows the A2DP sink component to
stream audio.
Add a TODO pointing to the mediasession migration so that this
capability can be offered hermetically once the component is migrated to
CFv2.

Don't include the test by default as it requires hardware to pass.

Bug: 74812
Test: fx test bt-a2dp-loopback-test
Test: Manual, fx test bt-a2dp-loopback-test on sherlock and verify test
streams for 10seconds with successful exit.

Change-Id: Ia5bcd2432c54c05d5b1bb067f3cf089895f8f790
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706486
Reviewed-by: Ankur Mittal <anmittal@google.com>
Commit-Queue: Ani Ramakrishnan <aniramakri@google.com>