[ffx][config] Clean up edge-case lockfile failures
If we fail to write out details to the lockfile (which should be pretty
rare if we managed to open the file to begin with, but seems like it
might be happening sometimes), we should remove the incorrect lockfile
before returning Err.

We should also flush the file to disk immediately and not return until
the file is flushed.

This hopefully increases the reliability of the lockfiles, and produces
useful output more of the time when something goes wrong. To help with
that too, let's output the current process' pid in the error as well so
it's a little more visible if a process is deadlocking itself.

Change-Id: Ib7aa4caf3cb8b5f7f000ee39f0054473f259ed95
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708533
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>