[virtualization] Enable the rust virtio-gpu/input devices by default
I'm in the process of enabling the integration tests for these
components. Since the older c++ ones had many tests disabled and they
rely on different APIs (gfx vs flatland), it will be simpler to enable
these devices along with the tests.

The risk should be limited since these devices do not see heavy use
(they're unused in workstation) and have good component level testing.
Compared to the c++ devices, we do not yet support mouse or cursor
commands but these should not be a large amount of work to re-implement
on top of flatland.

Bug: 102870, 104229
Change-Id: I58d98b7da1f2a6229c151b9a462c029afe77029a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706914
Reviewed-by: Alexander Radchenko <alerad@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>