[fxtest] Skip loading package repository for host tests
The package repository is loaded to add package hashes to URLs for
tests that run as components on the target device. This ensures that
the target device donwloads the latest version of the test before
running it.

Host tests are not components and do not need hashes to run the latest
version. Instead, fx test expects to find the latest version of the
binary in the output directory. In products such as bringup, there are
no components and the package repository is not generated.

This change lazily loads package repository, so that no attempt to load
it is made unless there is a test that requires it. Note that in all
cases, tests are still rebuilt if code has changed.

Testing: new unit tests.
Manually verify fx test:
 * 'fx test zither' (host test) succeeds after clean-build on bringup
   and executes the latest version of the test
 * 'fx run_test_suite' (target test) on core adds package hash and
   executes the latest version of the test
 * 'fx run_test_suite --no-use-package-hash' on core does not add the
   package hash

Fixed: 105846

Change-Id: Ib9891a7ffac584925217c138c8ddf310c6d35d1e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708531
Commit-Queue: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Gary Boone <gboone@google.com>
Fuchsia-Auto-Submit: Satsuki Ueno <satsukiu@google.com>