[microbenchmarks] Make HandleValid test's set of metric names deterministic
For the allowlist being added for performance metric names, it's
desirable for the set of metric names to be deterministic.

Before this change, the set of metric names would vary based on the
number of CPUs on the target system.  Address that by always testing
two thread counts (one thread, and one thread per CPU).

Note that the NUCs we test on only have two CPUs anyway (with
hyperthreading disabled, which is how we run the performance tests).

Bug: 105202
Change-Id: I570994c068ed23c3ffb6d68c9cd27b2a39492bb1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709286
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>