[kernel][page_cache] Allow waitable allocations to fill the cache.
Correct a bug in fxrev.dev/582047 the sent waitable page allocations
directly to the PMM instead of allowing them to fill the cache,
effectively bypassing the cache for all user VMOs. The bug does not
introduce a regression relative to behavior prior to 582047, however is
does defeat the intended optimization. This CL restores the intended
optimization.

Bug: b/234776870
Change-Id: I8c6e4807e07e8f9f33eb2adf895d253ef65f5859
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709011
Commit-Queue: Corey Tabaka <eieio@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>