[examples][fuzzing] Clean up build files for v2 migration
The fuzzer GN templates will switch to supporting component framework
v2 in an upcoming change. The templates are designed to be almost a
drop-in replacement, with component details being generated
automatically for fuzzers defined using the older v1 templates. This CL
cleans up two areas of the example fuzzers that are not completely
automatic:
  * Renamed fuzzers (via output_name) will need an explicit manifest, so
    the OOM fuzzer has been changed.
  * //build/components.gni is no longer transitively included.

Change-Id: Iad62c118c10d7b88832f6e41eb00d4190290853d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706884
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Cameron Finucane <eep@google.com>
Reviewed-by: Shai Barack <shayba@google.com>