[diagnostics] Rename triage lib to avoid conflict
Right now the existence of a triage library and a triage binary causes
generated cargo files to have a broken dependency tree in some cases
(where they depend on both the executable library and the lib library).
This renames the triage lib so it doesn't get confused with the binary
target anymore.

Found this while attempting to gen-cargo the ffx target, but the
dependency that creates this problem is in triage_lib_test.

Change-Id: I02cdb7a9d1a375ef289ee748a70d9fddaf2b6245
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709284
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>