[ffx] Add a console command to ffx emu
This CL adds a new command to the ffx emu tool set named "console". This
tool exposes the QEMU command consoles to the user after launch time by
connecting to them over Linux sockets that must be set up at launch
time. At this time, this is a simple send/receive tool; it has no
knowledge of what the messages mean, no history, and no support for
advanced command-line functionality like colors. It's purely
experimental, for the sake of exploring what the consoles can do and how
we can interact with them. To enable this tool, you must add the
"emu.sockets.enabled" flag to your ffx config.

The choice of which console to connect to is accomplished via the
"--serial", "--command", and "--machine" flags respectively. These are
mutually exclusive flags, and the user may specify one of these flags
or the tool will default to "serial".

Bug: 98236

Change-Id: Ib1ba24b089c1ad753697f4fe2c26edb17c430b11
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684144
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>