[local_storage] Change logged metric type to Integer for two metrics.
These two metrics are incorrectly being logged as OCCURRENCE metric
type, when they are INTEGER metric types in the Cobalt registry.
This was caused by the migration from Cobalt 1.1, due to the
replacement metric being of type INTEGER when the original 1.0
metric was of type EVENT_COUNT (which in Lapis is an Occurrence
metric type). See the bug for more details on this issue.

Bug: 106179
Change-Id: I887273c372cc5c738f035285c61110c037e1ce47
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709768
Reviewed-by: Brandon Castellano <bcastell@google.com>
Reviewed-by: Chris Phoenix <cphoenix@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Cameron Dale <camrdale@google.com>