[camera] Fix no frames after unmuting
This CL fixes a problem with clients receiving no frames after unmuting.

On some occasions, the image sensor does not appear to fully come out of
reset when unmuting. I2C transactions fail, and no frames are produced
by the image sensor.

This CL contains a work around for this problem. Unmuting leads to a
mode change. Before processing the mode change, which will prepare the
sensor for subsequent streaming, the sensor driver will now read the
device ID from the sensor to test I2C communication. If this fails, the
sensor will be cycled through reset and the test repeated. Should this
again result in an I2C failure, the reset process will be repeated up to
MAX_RESET_ATTEMPTS times (currently set to 4) before reporting failure.

This has been tested on a device and shown to address the problem.

Bug 239845798

Change-Id: Ifbc8fe033a9782bd7e5de565e0a1077bc4b24ca1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709346
Reviewed-by: Ernest Hua <ernesthua@google.com>
Commit-Queue: John Sasinowski <jsasinowski@google.com>