[settings] Update config tests to verify light group names have unique hashes
Light group names in the light hardware configuration are expected to be
unique and also to have unique hashes in order for the light setting to
function properly. This CL adds a test that panics if two light group
names hash to the same value.

Fixed: 103000
Test: fx test sample-setui-config-test
      locally tested with two identical light group names and verified
      the test fails. Unfortunately I don't know of two strings that
      hash to the same value in Rust.
Change-Id: I3b508e9220f09e17a49b326343134384204fee40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706592
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>
Reviewed-by: Sijie Chen <sijiec@google.com>
Commit-Queue: William Xiao <wxyz@google.com>