[ffx][target] Format IpV6 Address in TargetAddr
Update Display for TargetAddr to write the IPV6 addresses with
surrounding square brackets "[]".

Also if the address is a local address and has a scope, then it will translate
the scope id to its associated interface name (if possible) and append
that to the address.

Bug: 104635
Change-Id: Ie6bd425887daba7c07852ac1715e82fbf4e5c8fb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/707970
Fuchsia-Auto-Submit: Colin Nelson <colnnelson@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>