[pbm] Support PBMs with correct relative paths
bundle_fetcher now supports PBMs with the new relative paths, and falls
back to the old format if those files are not found. We have to complete
this in a soft-transition, because bundle_fetcher is pinned to a
specific version in infra.

Bug: 106117
Change-Id: I7bd0dcab64507264dbe58e0a04d8d363fc6942df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709886
Reviewed-by: Anthony Fandrianto <atyfto@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>