[fxfs] Manage volumes explicitly on the client
Lifts the responsibility of creating the default volume to the client.
For now, we bake it into fs_management and fshost to always do this, so
this is generally a pure refactoring in terms of external behaviours.

That said, this involves a sweeping change to fs_management and all
clients thereof.  This is centered around the introduction of volume
management to fs_management, via:
- StartedMultiVolumeFilesystem in C++, and
- ServingMultiVoluemFilesystem in Rust.

A few other changes are made:
- migrate_root is no longer used and can be removed in a followup CL.
- Fxfs no longer implements request queueing, this is done by
  fshost anyways.
- Some minor adjustments are made to the export root layout, e.g.
  volumes export a svc/ directory.
- A number of minor bugs were fixed, e.g. swallowing errors during
  start (cf. fxrev.dev/702516) and the fs-management minfs tests
  not actually launching a componentized minfs when they say they do.

Bug: 93066
Bug: 104313

Change-Id: If6bfbe7f384299211febf9779ccabed1e1972d25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700417
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Eric Stone <ecstone@google.com>
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Size-Review: Gopi Krishna Chitluri <gopichitluri@google.com>