[fidlgen][cpp] Introduce DeclDepGraph
This change introduces an abstraction for computing the dependency
graph among a library's declarations. It will allow use to remove
`declaration_order` from the IR - and, as a demonstration of that, we
migrate the C++ bindings over to relying upon this instead in the
following change.

We also fix a minor fidlc bug that was generating the wrong JSON
"maybe_size" element for layout members (on which this CL introduces a
dependency).

Bug: 7660, 51002
Change-Id: I6b31d40bbaab8f22ba245a55295837ff7ff78dab
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/703884
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>