[netstack3] Combine multicast interface selectors
Use a single type to represent the possible ways to select a multicast
interface when making a membership change. This makes it impossible to
represent the case where an interface ID and local address are provided,
pushing the responsibility of deciding what that means up to the caller.

Bug: 39479
Change-Id: Ifb0a862225a28b17ecef3f850761db82b6b073d7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710224
Commit-Queue: Alex Konradi <akonradi@google.com>
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>