[virtualization] Use UITestManager for virtualization-input-tests
This virtualization-input-test used to rely on a 'FakeScenic'
implementation. At some point these tests had been disabled and haven't
been building or running for some time.

This CL removes this 'FakeScenic' functionality and instead integrates
with the UITestManager to provide a hermetic UI stack. There has been
some changes in the UI and Input protocols and so this test is currently
only written to support the rust GPU and Input devices.

Bug: 104229, 64348
Test: fx test virtualization-input-tests
Multiply: virtualization-input-tests
Change-Id: I2a0b9f565e1400ce6c659429638b0796696df9e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/706623
Reviewed-by: Alexander Radchenko <alerad@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Alexander Brusher <abrusher@google.com>
Reviewed-by: David Hastings <dahastin@google.com>