[virtualization] Re-enable virtio-gpu tests
These integration tests had been disabled and had acquired some bitrot
because they had not been building in CI for some time. This CL revives
the binary but keeps many of the pre-existing tests as disabled. The
tests are now configured to run in a UITestManager environment that has
access to a hermetic graphics stack.

We introduce a new `virtio_gpu_test_util` binary that runs in a
(currently linux-only) guest and uses the linux Framebuffer API to
detect display geometry. We add a new test that simply compares that the
guest display geometry matches that of the Fuchsia view in the hermetic
UI stack.

Bug: 102870
Test: fx test virtualization-gpu-tests --count 100
Multiply: virtualization-gpu-tests
Change-Id: Ia71f0c1bd6a9e7d80a2474668b99d8795b00fafa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708566
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: David Hastings <dahastin@google.com>