[zxdb][DAP] Initialize Request returns failure if not connected
This changes the DAP handler for the InitializeRequest message
so that if the debugger is not connected to the device, an error
is returned indicating the initialization failed and the debug
session should not continue.

The error message from the connection attempt is stored in the
session and set when the connection attempt is completed.

Fixed: 105906
Change-Id: Id38fe1ddce0ca9f4334f3a9917efd2e342d82c49
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709703
Reviewed-by: Dangyi Liu <dangyi@google.com>
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>