[cmc] Remove restricted feature for structured config.
We have completed all of our plans for structured config changes
that would be difficult to soft-migrate, so it's now pretty safe
for OOT integrations to be built.

This change does not on its own make it possible for OOT usage,
but the build integrations we will now write will no longer need
to opt-in to an unstable feature.

Bug: 95368
Bug: 101970
Change-Id: I0c0c07b4e80ec116f8aaef3aff92bdb287776e74
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710404
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Adam Perry <adamperry@google.com>