[sysmem] sysmem-unittest migration fixup
We were running sysmem-bootfs-test but not sysmem-test,
so the migration of sysmem-test needed a couple tweaks.

To avoid this sort of regression in future, go ahead
and run sysmem-test as well.  The actual tests are the
same, but we do want to make sure the tests pass in both
environments.

Change-Id: I06222459e5c2d7737cdabf9f7a4dc357aa4cee2e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709244
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: Shai Barack <shayba@google.com>