[virtualization] Add ZirconGuest to gpu tests.
This enables a simple test to detect displays for the zircon guest. To
support zircon, we also modify the behavior of the rust virtio-gpu
device to wait for the backing view to be created before processing
queues. This is needed because the zircon driver does not yet implement
the config-changed interrupt and will be unable to detect display
hot-plug or resize.

Bug: 102870, 64348
Test: fx test virtualization-gpu-tests --count 100
Multiply: virtualization-gpu-tests
Change-Id: I65c7f644d9b0128bf27f92b05f42f73fdd80ab56
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/708743
Reviewed-by: David Hastings <dahastin@google.com>
Reviewed-by: Alexander Radchenko <alerad@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>