[fidl][hlcpp] unknown interactions client.
Update HLCPP client to be able to handle unknown interaction responses.
My approach to doing this is just to expose the result union as a
client-usable type, so this actually requires fairly minimal changes,
and most of this change is updating the HLCPP unit tests to cover the
entire suite of client tests copied from the tests for the new C++
bindings.

Bug: 88366

Change-Id: Ieaaf8f4bafffc82f1e403b32250eba1cf29a4429
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705107
Commit-Queue: Zachary Stewart <zstewart@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>