[scsi] Check write_cache_enabled bit before flush
Issuing a SYNCHRONIZE command to a disk that doesn't enable a write
cache leads to an error. Therefore, we have to verify that the write
cache is enabled before flushing. In a general solution, we'd have to
check the bit before every single flush, but since we don't support MODE
SELECT commands to change the bit, we can just read it on disk
initialization and cache the value.

Fixed: 86850
Change-Id: Id5a48d6d358fd9e450188948deb794f0719f70c0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710162
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Venkatesh Srinivas <venkateshs@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Braden Kell <bradenkell@google.com>