[bt][audio-output] Add initial inspect
Add inspect support to SoftPcmOutput.
Refactor so that the audio stream processing is started
automatically when the stream begins to be polled, and the task is
dropped instead of detached.

This also means that if AudioCore disconnects from the RingBuffer or
StreamConfig, the audio stream will stop accordingly.

Bug: 73682
Test: inspect changes only + refactor, confirm with snapshot
Test: fx test fuchsia-audio-device-output-tests
Change-Id: Ie7f909b06a5487652022c2a56acdf257d3133452
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/664685
Commit-Queue: Marie Janssen <jamuraa@google.com>
Reviewed-by: Dayeong Lee <dayeonglee@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>