[component_manager] Fix flake in rights_integration_test
This test was opening a directory capability that was expected
to fail due to a rights issue. It then tried to open a file in
the directory and write to it.

These fuchsia.io calls got pipelined.

Unfortunately a failure could occur either when the file was
opened (because CM failed to route the capability and closed the
connection early) or while attempting to write (because the test
would wait at the end of the pipelined calls for a failure).

This test did not expect the opening of the file to fail and
so the test flaked, depending on how fast CM went through the
routing code.

This test now uses the DESCRIBE flag to open the file (the
fuchsia_fs::directory::open_file call does this).

Also completed the routes for fuchsia.logger.LogSink which
helped debug this test.

Fixed: 103991
Test: Ran this test ~3300 times
Multiply: rights_integration_test
Change-Id: I0790fad0863f7c6530acc0c1fc06bb57d183333b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/710942
Commit-Queue: Xyan Bhatnagar <xbhatnag@google.com>
Reviewed-by: Yaneury Fermin <yaneury@google.com>
Fuchsia-Auto-Submit: Xyan Bhatnagar <xbhatnag@google.com>