[driver manager] Integrate DeviceGroupManager to DFv1
Replace the old device group logic in DFv1 with the
new one with DeviceGroupManager. Coordinator now owns
a DeviceGroupManager and sends device group requests to
it.

Coordinator is modified so that it inherits and
implements CompositeManagerBridge. This CL adds a
DeviceGroupV1 class, which subclasses DeviceGroup and
handles a CompositeDevice.

Bug: 104950
Tests: Updated misc_tests
Change-Id: I12b761a753a0a30c7fd072f7c8f11edfabf9d251
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/705483
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>