[fidl][rust] Support allowlisting in backend
This change adds support for feature-based allowlists, with membership
determined by library name, in fidlgen_rust. This is necessary for the
large messages prototype, as upcoming changes will build out this
prototype, but only enable it for a small set of FIDL libraries.

Being included in an allowlist triggers a comment being added in the
resulting Rust code, like `// allowlisted_by = my_allowlist`. This CL
thus also includes a new, empty golden FIDL library to test that this
occurs as expected.

Test: fx regen-goldens; fx test fidlgen_rust_lib_tests
Bug: 106641
Change-Id: I5101972d7fd1da80822267e62ea24fd0f078599d