[fidl][cpp] Publicize DecodedValue
DecodedValue is a utility wrapper to close handles inside a decoded
object tree. This will be used in the next CLs as part of a standalone
API to encode and decode FIDL objects (FIDL-at-rest).

This CL moves it out of the "internal" namespace and adds a bit of API
polish. For example, |value| is typically understood to return a
reference type.

Bug: 82681

Change-Id: I32598915052a7720de4ff2414045df54a9574adb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/709462
Reviewed-by: Zachary Stewart <zstewart@google.com>
Commit-Queue: Yifei Teng <yifeit@google.com>