[fdf] Remove tools-drivers and tools-system
Remove  "//bundles:tools-system" and "//bundles:tools-drivers".

Originally, tools-drivers was to replace tools-system in order to avoid
driver packages being inserted in the system packaget set. However, it
turns out that the only target in tools-system, a bluetooth hci emulator
driver package, is no longer needed to be loaded into Fuchsia and so the
tools-system can just be removed and not replaced.

Bug: 69738
Change-Id: I366cedce61b0ce2cbc607cac17ec9f9a4b622931
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699942
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Nick Eaton <nickeaton@google.com>