[kernel][vm] Do not dirty forked zero page on commit
When a zero page marker is replaced with a forked zero page in a
pager-backed VMO, only mark it dirty if we were looking up the page with
mark_dirty set to DirtyAllPagesOnWrite. This ensures we do not dirty
forked zero pages during a commit, which passes the VMM_PF_FLAG_WRITE so
that it can fork pages if required, but does not indicate
DirtyAllPagesOnWrite.

Test: core-pager-writeback
Bug: 63989
Run-All-Tests: true
Change-Id: I2cf3d635476ee6fb9d2b4810907bada10a9f7aa4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/711243
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>